// frontend/src/app/models/odrzavanje.ts
export class Odrzavanje
{
  idO: number = 0;
  status: string = '';
  datumZavrsetka: string = '';
  vremeZavrsetka: string = '';
  idZ: number = 0;
  idF: number = 0;
  dekorater: string = '';

  nazivFirme: string = '';
  brBazena: number=0;
  brFontana: number = 0;
  tipBaste: string = "";
  kvBazen: number = 0;
  kvFontana: number = 0;
  idRaspored: number = 0;
}